moved metadata route into source_app/api.py
