+++ securedrop/source.py
@@ -1,27 +1,9 @@
 # -*- coding: utf-8 -*-
-from flask import render_template, make_response
-
 import config
-import json
-import version
+
 from source_app import create_app
 
-import logging
-# This module's logger is explicitly labeled so the correct logger is used,
-# even when this is run from the command line (e.g. during development)
-log = logging.getLogger('source')
-
 app = create_app()
-
-
-@app.route('/metadata')
-def metadata():
-    meta = {'gpg_fpr': config.JOURNALIST_KEY,
-            'sd_version': version.__version__,
-            }
-    resp = make_response(json.dumps(meta))
-    resp.headers['Content-Type'] = 'application/json'
-    return resp
 
 
 if __name__ == "__main__":  # pragma: no cover
