+++ securedrop/source_app/__init__.py
@@ -16,7 +16,7 @@ import version
 
 from db import Source, db_session
 from request_that_secures_file_uploads import RequestThatSecuresFileUploads
-from source_app import main, info
+from source_app import main, info, api
 from source_app.decorators import ignore_static
 from source_app.utils import logged_in
 
@@ -55,8 +55,8 @@ def create_app(config=None):
     app.jinja_env.filters['nl2br'] = evalcontextfilter(template_filters.nl2br)
     app.jinja_env.filters['filesizeformat'] = template_filters.filesizeformat
 
-    for blueprint in [main.make_blueprint(), info.make_blueprint(config)]:
-        app.register_blueprint(blueprint)
+    for module in [main, info, api]:
+        app.register_blueprint(module.make_blueprint(config))
 
     @app.before_request
     @ignore_static
